requirements.txt
click>=8.0.0
GitPython>=3.1.0
pyyaml>=6.0
pathspec>=0.10.0 